platformqctl/requirements.txt
click
ruamel.yaml
Jinja2
httpx
pulsar-client
trino
tabulate 